ui/views/logs.py
"""
logs.py: Runtime + error logs view.
"""
from flask import jsonify

def logs_view():
    # TODO: Show logs from logs/orchestrator.log
    return jsonify({"status": "ok", "message": "Logs placeholder"})
